requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coloredlogs==15.0
cssselect==1.1.0
Deprecated==1.2.12
h11==0.9.0
howdoi==2.0.15
httpcore==0.11.1
httptools==0.2.0
httpx==0.15.4
humanfriendly==9.1
idna==2.10
keep==2.10.1
lxml==4.6.3
multidict==5.1.0
newsapi-python==0.2.6
pycparser==2.20
PyGithub==1.55
Pygments==2.8.1
PyJWT==2.1.0
PyNaCl==1.4.0
pyquery==1.4.3
pyreadline==2.1
rasa-sdk==2.5.0
requests==2.25.1
rfc3986==1.4.0
sanic==20.12.3
Sanic-Cors==0.10.0.post3
Sanic-Plugins-Framework==0.9.5
six==1.15.0
sniffio==1.2.0
terminaltables==3.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
websockets==8.1
wrapt==1.12.1
